// src/app/models/menujour.ts
export class MenuJour {
  id: number;
  prixdujour?: string;
  menuviande?: string;
  menupoisson?: string;
  prixformule?: string;
  prixmenu?: string;
  libellesugg1?: string;
  prixsugg1?: string;
  descsugg1?: string;

  libellesugg2?: string;
  prixsugg2?: string;
  descsugg2?: string;

  libellesugg3?: string;
  prixsugg3?: string;
  descsugg3?: string;

}